+++ src/dashboardWebView/DashboardMessage.ts
@@ -25,4 +25,7 @@ export enum DashboardMessage {
   getDataEntries = 'getDataEntries',
   putDataEntries = 'putDataEntries',
   sendTelemetry = 'sendTelemetry',
+  insertSnippet = 'insertSnippet',
+  addSnippet = 'addSnippet',
+  updateSnippet = 'updateSnippet',
 }