+++ src/dashboardWebView/components/Dashboard.tsx
@@ -9,6 +9,7 @@ import { Contents } from './Contents/Contents';
 import { Media } from './Media/Media';
 import { NavigationType } from '../models';
 import { DataView } from './DataView';
+import { Snippets } from './SnippetsView/Snippets';
 
 export interface IDashboardProps {
   showWelcome: boolean;
@@ -31,6 +32,14 @@ export const Dashboard: React.FunctionComponent<IDashboardProps> = ({showWelcome
     return <WelcomeScreen settings={settings} />;
   }
 
+  if (view === NavigationType.Snippets) {
+    return (
+      <main className={`h-full w-full`}>
+        <Snippets />
+      </main>
+    );
+  } 
+  
   if (view === NavigationType.Media) {
     return (
       <main className={`h-full w-full`}>