+++ src/dashboardWebView/models/NavigationType.ts
@@ -2,4 +2,5 @@ export enum NavigationType {
   Contents = "contents",
   Media = "media",
   Data = "data",
+  Snippets = "snippets",
 }