+++ src/helpers/Extension.ts
@@ -2,7 +2,7 @@ import { existsSync, renameSync } from "fs";
 import { basename, join } from "path";
 import { extensions, Uri, ExtensionContext, window, workspace, commands, ExtensionMode, DiagnosticCollection, languages } from "vscode";
 import { Folders, WORKSPACE_PLACEHOLDER } from "../commands/Folders";
-import { EXTENSION_NAME, GITHUB_LINK, SETTINGS_CONTENT_FOLDERS, SETTINGS_CONTENT_PAGE_FOLDERS, SETTING_DATE_FIELD, SETTING_MODIFIED_FIELD, SETTING_SEO_DESCRIPTION_FIELD, SETTING_TAXONOMY_CONTENT_TYPES, DEFAULT_CONTENT_TYPE_NAME, EXTENSION_BETA_ID, EXTENSION_ID, ExtensionState, DefaultFields, LocalStore, SETTING_TEMPLATES_FOLDER } from "../constants";
+import { EXTENSION_NAME, GITHUB_LINK, SETTING_CONTENT_FOLDERS, SETTING_CONTENT_PAGE_FOLDERS, SETTING_DATE_FIELD, SETTING_MODIFIED_FIELD, SETTING_SEO_DESCRIPTION_FIELD, SETTING_TAXONOMY_CONTENT_TYPES, DEFAULT_CONTENT_TYPE_NAME, EXTENSION_BETA_ID, EXTENSION_ID, ExtensionState, DefaultFields, LocalStore, SETTING_TEMPLATES_FOLDER } from "../constants";
 import { ContentType } from "../models";
 import { Notifications } from "./Notifications";
 import { parseWinPath } from "./parseWinPath";
@@ -137,7 +137,7 @@ export class Extension {
 
     // Migration to version 3.1.0
     if (major < 3 || (major === 3 && minor < 1)) {
-      const folders = Settings.get<any>(SETTINGS_CONTENT_FOLDERS);
+      const folders = Settings.get<any>(SETTING_CONTENT_FOLDERS);
       if (folders && folders.length > 0) {
         const workspace = Folders.getWorkspaceFolder();
         const projectFolder = basename(workspace?.fsPath || "");
@@ -147,7 +147,7 @@ export class Extension {
           path: `${WORKSPACE_PLACEHOLDER}${folder.fsPath.split(projectFolder).slice(1).join('')}`.split('\\').join('/')
         }));
 
-        await Settings.update(SETTINGS_CONTENT_PAGE_FOLDERS, paths);
+        await Settings.update(SETTING_CONTENT_PAGE_FOLDERS, paths);
       }
     }